+++ src/styles.ts
@@ -1,4 +1,4 @@
-import { createStyles, makeStyles, Theme } from "@material-ui/core";
+import { createStyles, makeStyles } from "@material-ui/core";
 
 
 export const formStyles = makeStyles(theme => createStyles({