+++ CHANGELOG.md
@@ -1,7 +1,15 @@
 # Change Log
 
 
-## [0.14.0] - 2020-09-30
+## [0.14.2] - 2020-10-27
+### Added
+- `columnWidth` field in properties to indicate the width of columns
+
+### Changed
+- Fix for wrong subcollections url.
+
+
+## [0.14.0] - 2020-10-25
 ### Added
 - Lateral panel for having entities info in context
 - Revamped collection table for allow infinite scrolling and enhanced performance